RAG project/requirements.txt
langchain
ipynb
unstructured
libmagic
python-magic
python-magic-bin
faiss-cpu
sentence-transformers
streamlit
python-dotenv
openai
tiktoken
dill
